Sloppy but functional desktop top nav
