+++ src/components/dev-hub/global-nav.js
@@ -1,17 +1,16 @@
-import React, { useState } from 'react';
+import React from 'react';
 import dlv from 'dlv';
 import styled from '@emotion/styled';
 import { graphql, useStaticQuery } from 'gatsby';
 import DevLeafDesktop from './icons/mdb-dev-leaf-desktop';
 import DevLeafMobile from './icons/mdb-dev-leaf-mobile';
 import Link from '../Link';
-import { P, P3 } from './text';
-import { fontSize, layer, lineHeight, screenSize, size } from './theme';
+import { fontSize, lineHeight, screenSize, size } from './theme';
 import useMedia from '~hooks/use-media';
+import NavItem from './nav-item';
 
 // nav height is 58px: 24px line height + 2 * 17px vertical padding
 const LINK_VERTICAL_PADDING = '17px';
-const SUBITEM_MAX_WIDTH = '364px';
 
 const GlobalNav = styled('nav')`
     background-color: ${({ theme }) => theme.colorMap.greyDarkThree};
@@ -73,42 +72,6 @@ const HomeLink = styled(NavLink)`
     }
 `;
 
-const NavListHeader = styled(NavLink)`
-    position: relative;
-`;
-const NavItemList = styled('div')``;
-const NavItemSublist = styled('ul')`
-    display: ${({ isExpanded }) => (isExpanded ? 'block' : 'none')};
-    list-style: none;
-    margin-top: 0px;
-    padding-top: 1px;
-    border-top: 1px solid transparent;
-    max-width: ${SUBITEM_MAX_WIDTH};
-    padding-left: 0;
-    position: absolute;
-    z-index: ${layer.front};
-    > li {
-        background-color: ${({ theme }) => theme.colorMap.greyDarkThree};
-        margin: 0;
-        padding: ${size.medium} ${size.large};
-        &:hover,
-        &[aria-current='page'] {
-            background-color: #2c3d47;
-        }
-        :not(:last-of-type) {
-            box-shadow: 0px 1px 0px #3d4f58;
-        }
-    }
-`;
-const SubItemLink = styled(NavLink)`
-    height: 100%;
-    width: 100%;
-    padding: 0;
-`;
-const SubItemDescriptionText = styled(P3)`
-    color: ${({ theme }) => theme.colorMap.greyLightTwo};
-`;
-
 const topNavItems = graphql`
     query TopNavItems {
         strapiTopNav {
@@ -118,44 +81,6 @@ const topNavItems = graphql`
         }
     }
 `;
-
-const NavItemSubItem = ({ subitem }) => (
-    <SubItemLink to={subitem.url}>
-        <div>
-            <P>{subitem.name}</P>
-            <SubItemDescriptionText collapse>
-                {subitem.description}
-            </SubItemDescriptionText>
-        </div>
-    </SubItemLink>
-);
-
-const NavItem = ({ item }) => {
-    const [isExpanded, setIsExpanded] = useState(false);
-    const hasSubMenu = !!item.subitems.length;
-    if (hasSubMenu) {
-        const NavListHeaderDiv = NavListHeader.withComponent('div');
-        return (
-            <NavItemList
-                onBlur={() => setIsExpanded(false)}
-                onMouseLeave={() => setIsExpanded(false)}
-                onMouseEnter={() => setIsExpanded(true)}
-                onFocus={() => setIsExpanded(true)}
-                tabIndex="0"
-            >
-                <NavListHeaderDiv>{item.name}</NavListHeaderDiv>
-                <NavItemSublist isExpanded={isExpanded} tabIndex="0">
-                    {item.subitems.map(subitem => (
-                        <li>
-                            <NavItemSubItem tabIndex="0" subitem={subitem} />
-                        </li>
-                    ))}
-                </NavItemSublist>
-            </NavItemList>
-        );
-    }
-    return <NavLink to={item.url}>{item.name}</NavLink>;
-};
 
 export default () => {
     const data = useStaticQuery(topNavItems);
